Feat: retry opening files that should exist already
Except for the first construction, `FileDisk` instances are only reading
or writing files that should have already been created. If the file has
magically disappeared, the user probably moved it elsewhere to free disk
space, and will likely move it back (or quit plotting manually if they
can't). This change copies the "wait 5 minutes and retry" behaviour of
read, write, copy, and rename failures in `FileDisk` and `DiskPlotter`
to `FileDisk`'s `Open` member.