Reopen files upon read or write failure
If an attempt to read or write a temp file fails, close the
file before sleeping, then re-open it after sleeping.

This improves robustness when plotting with removable drives.
Previously, if a drive is unmounted/remounted during the plotting
process, the open files never recover and you get stuck in a loop
of:

> Only read 0 of <size> bytes at offset <offset> from "<filename>"
> with length <length>. Error 1. Retrying in five minutes.

This change closes and re-opens files and can recover from (some?)
transient read and write problems. Specifically, if you have a
removable drive, and you unplug it during phase 2, it will recover
and continue when you plug the drive back in.