Always on error reset position/pointer
If an error occurs, the resulting value of the file-position indicator for the stream is unspecified. https://pubs.opengroup.org/onlinepubs/007904975/functions/fwrite.html

And in the code above if error occurs with 0 bytes written (full disk) it will not reset the pointer (writePos will still be equal to begin), however it need to be reseted. 

Otherwise this causes https://github.com/Chia-Network/chiapos/issues/234 - in phase3, when this bucket is read, it goes into endless loop.