tests: Move `(De)serialization` test case down in hierarchy
This is to work around an issue in the `blake3` source code namely the 
following function having a "init phase" which isn't thread safe 
(Assigning the global variable `g_cpu_features` if its undefined yet):

https://github.com/BLAKE3-team/BLAKE3/blob/4056af6d7ffdf4d13bb776b7ea1db2a6b52d4d75/c/blake3_dispatch.c#L85 

Since it's not thread safe it lead to data race here prior to this 
commit because 24e3057e05fcaeae040b115d49d5452a1cfa064e introduced the 
`(De)Serialization` test as first test in the file and it triggeres plot 
creation with 2 phase1 threads which trigger `get_cpu_features` and in 
this case concurrently. The workaround here is to move the new test 
below the `F Functions` test which is the first test in the file which 
is calling `get_cpu_features` from a single thread. This single thread 
runs the "init phase" and since this is only required once the follow up 
tests are as fine like they were before on `main`.